Disable StopAutoDraft Plugin Universally
This PR changes the logic of the StopAutoDraft plugin to simply set the new `Vanilla.Drafts.Autosave` config value to `false`. This setting is used by the new settings option on the `vanilla/settings/posting` page. Once the config value is set, the addon disables itself. The PR also hides the plugin, so it cannot be found subsequently.